backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.32
alembic==1.13.2

python-dotenv==1.0.1
pydantic==2.10.4
requests==2.32.3

cohere==5.5.8
tavily-python==0.3.4

# Orchestration (requires Python 3.12 for prebuilt wheels on Windows)
langchain==0.2.16
langchain-cohere==0.2.1
langchain-community==0.2.16
langgraph==0.2.15

# MCP (optional)
mcp==1.2.0
anyio==4.6.2.post1
websockets==12.0
 
# Note: Postgres driver (psycopg2-binary) is optional and can cause build issues on Windows.
# If/when you switch from SQLite to Neon Postgres, install:
#   pip install -r requirements-postgres.txt
